requirements.txt
opencv-python
pycryptodome
numpy
scipy
PyWavelets
